[Fix] query문 위치 변경
